Fix for "cannot create CANCEL" issue
Issue description:
The transaction termination timeout of a PRACK transaction (5s after
receiving the 200 OK for PRACK) caused the original INVITE transaction
that created the dialog to be cleaned up. After this point, it was
impossible to create a CANCEL for that INVITE, as the transaction could
not be found anymore.

Fix description:
- don't overwrite an existing session creating transaction/dialog by
subsequent transactions
- do not update SipSessionImpl.cseq for outgoing, locally generated
requests, because cseq is only used for CSeq validation of incoming
requests
- do not falsely report 'isAckReceived:true' for the initial INVITE if
one hasn't arrived yet
- do not call isAckReceived for dialogs started as UAC (not applicable)
- only clean the INVITE transaction if the session is already confirmed