Fix for resp100rel.isCommitted()==false after PRACK sent
Since commit 4aae34316, the INVITE transaction does not terminate any more after the PRACK's 200 OK arrives (since that is a completely different transaction). This broke the logic in SipServletResponseImpl.isCommitted(), which I believe was wrong in the first place. Corrected to use the already present isPrackGenerated field instead of the transaction state.