Fixed encoding of header name/value in SipURI
getHeader() was decoding the value returned by the underlying javax.sip.address.SipURI, but
setHeader() was not passing down a properly encoded value. The latter is now fixed, and the
header name is also encoded properly. Accordingly, the getHeaderNames() method was also
updated to present decoded names to the user.

Added a unit test case to check the correctness of the above behavior.